fix disk by distributing tests
