+++ tests/openvino/test_modeling.py
@@ -15,7 +15,6 @@
 import copy
 import gc
 import os
-import platform
 import tempfile
 import time
 import unittest
@@ -38,7 +37,6 @@ from sentence_transformers import SentenceTransformer
 from transformers import (
     AutoConfig,
     AutoFeatureExtractor,
-    AutoImageProcessor,
     AutoModel,
     AutoModelForAudioClassification,
     AutoModelForAudioFrameClassification,
@@ -48,12 +46,9 @@ from transformers import (
     AutoModelForImageClassification,
     AutoModelForMaskedLM,
     AutoModelForQuestionAnswering,
-    AutoModelForSeq2SeqLM,
     AutoModelForSequenceClassification,
-    AutoModelForSpeechSeq2Seq,
     AutoModelForTextToSpectrogram,
     AutoModelForTokenClassification,
-    AutoModelForVision2Seq,
     AutoModelForZeroShotImageClassification,
     AutoProcessor,
     AutoTokenizer,
@@ -69,9 +64,6 @@ from transformers.utils import http_user_agent
 from utils_tests import (
     MODEL_NAMES,
     TEST_IMAGE_URL,
-    get_num_sdpa,
-    mock_torch_cuda_is_available,
-    patch_awq_for_inference,
 )
 
 from optimum.exporters.openvino.model_patcher import patch_update_causal_mask
@@ -92,10 +84,8 @@ from optimum.intel import (
     OVModelForQuestionAnswering,
     OVModelForSeq2SeqLM,
     OVModelForSequenceClassification,
-    OVModelForSpeechSeq2Seq,
     OVModelForTextToSpeechSeq2Seq,
     OVModelForTokenClassification,
-    OVModelForVision2Seq,
     OVModelForVisualCausalLM,
     OVModelForZeroShotImageClassification,
     OVModelOpenCLIPForZeroShotImageClassification,
@@ -118,7 +108,6 @@ from optimum.intel.openvino.utils import (
     OV_VISION_EMBEDDINGS_MODEL_NAME,
     OV_VISION_ENCODER_MODEL_NAME,
     TemporaryDirectory,
-    _print_compiled_model_properties,
 )
 from optimum.intel.pipelines import pipeline as optimum_pipeline
 from optimum.intel.utils.import_utils import (
@@ -138,14 +127,9 @@ from optimum.utils import (
 from optimum.utils.testing_utils import require_diffusers
 
 
-TENSOR_ALIAS_TO_TYPE = {
-    "pt": torch.Tensor,
-    "np": np.ndarray,
-}
-
 SEED = 42
-
 F32_CONFIG = {"INFERENCE_PRECISION_HINT": "f32"}
+TENSOR_ALIAS_TO_TYPE = {"pt": torch.Tensor, "np": np.ndarray}
 
 
 class Timer(object):
@@ -1094,670 +1078,6 @@ class OVModelForFeatureExtractionIntegrationTest(unittest.TestCase):
             self.assertIn("Please use `OVSentenceTransformer`", str(context.exception))
 
 
-class OVModelForCausalLMIntegrationTest(unittest.TestCase):
-    SUPPORTED_ARCHITECTURES = (
-        "bart",
-        "baichuan2",
-        "baichuan2-13b",
-        "gpt_bigcode",
-        "blenderbot",
-        "blenderbot-small",
-        "bloom",
-        "chatglm",
-        "codegen",
-        "codegen2",
-        "gpt2",
-        "gptj",
-        "gpt_neo",
-        "gpt_neox",
-        "llama",
-        "marian",
-        "minicpm",
-        "mistral",
-        "mixtral",
-        "mpt",
-        "opt",
-        "pegasus",
-        "qwen",
-        "phi",
-        "internlm2",
-        "orion",
-        "falcon",
-        "falcon-40b",
-        "persimmon",
-        "biogpt",
-        "gpt_neox_japanese",
-        "xglm",
-        "aquila",
-        "aquila2",
-        "xverse",
-        "internlm",
-        "jais",
-        "chatglm4",
-        "decilm",
-    )
-
-    SUPPORTED_SSM_ARCHITECTURES = ("mamba", "falcon-mamba")
-    if is_transformers_version(">=", "4.43"):
-        SUPPORTED_ARCHITECTURES += SUPPORTED_SSM_ARCHITECTURES
-
-    if is_transformers_version(">=", "4.40.0"):
-        SUPPORTED_ARCHITECTURES += (
-            "gemma",
-            "olmo",
-            "stablelm",
-            "starcoder2",
-            "dbrx",
-            "cohere",
-            "qwen2",
-            "qwen2_moe",
-            "arctic",
-        )
-
-    if is_transformers_version(">=", "4.41.0"):
-        SUPPORTED_ARCHITECTURES += ("phi3",)
-
-    if is_transformers_version(">=", "4.43.0"):
-        SUPPORTED_ARCHITECTURES += ("gemma2", "exaone")
-
-    if is_transformers_version(">=", "4.44.0"):
-        SUPPORTED_ARCHITECTURES += ("granite", "granite-moe")
-
-    if is_transformers_version(">=", "4.46.0"):
-        SUPPORTED_ARCHITECTURES += ("glm", "mistral-nemo", "minicpm3", "phi3-moe")
-        # openvino 2025.0 required for disabling check_trace
-        if is_openvino_version(">=", "2025.0"):
-            SUPPORTED_ARCHITECTURES += ("deepseek",)
-
-        # gptq and awq install disabled for windows test environment
-        if platform.system() != "Windows":
-            SUPPORTED_ARCHITECTURES += ("opt_gptq",)
-
-        # autoawq install disabled for windows test environment
-        if is_openvino_version(">=", "2024.6.0") and platform.system() != "Windows":
-            SUPPORTED_ARCHITECTURES += ("mixtral_awq",)
-
-    if is_transformers_version(">", "4.49"):
-        SUPPORTED_ARCHITECTURES += ("gemma3_text",)
-
-    if is_transformers_version(">=", "4.51.0"):
-        SUPPORTED_ARCHITECTURES += ("qwen3", "qwen3_moe")
-
-    if is_transformers_version(">=", "4.51.3"):
-        SUPPORTED_ARCHITECTURES += ("glm4",)
-
-    if is_transformers_version(">=", "4.53.0"):
-        SUPPORTED_ARCHITECTURES += ("arcee",)
-
-    if is_transformers_version(">=", "4.54.0"):
-        SUPPORTED_ARCHITECTURES += ("ernie4_5",)
-
-    GENERATION_LENGTH = 100
-    REMOTE_CODE_MODELS = (
-        "chatglm",
-        "minicpm",
-        "baichuan2",
-        "baichuan2-13b",
-        "jais",
-        "qwen",
-        "internlm2",
-        "orion",
-        "aquila",
-        "aquila2",
-        "xverse",
-        "internlm",
-        "codegen2",
-        "arctic",
-        "chatglm4",
-        "exaone",
-        "decilm",
-        "minicpm3",
-        "deepseek",
-    )
-
-    EXPECTED_NUM_SDPA = {
-        "bart": 2,
-        "baichuan2": 2,
-        "baichuan2-13b": 2,
-        "gpt_bigcode": 5,
-        "blenderbot": 2,
-        "blenderbot-small": 2,
-        "bloom": 5,
-        "chatglm": 2,
-        "codegen": 5,
-        "codegen2": 2,
-        "gpt2": 5,
-        "gptj": 5,
-        "gpt_neo": 4,
-        "gpt_neox": 5,
-        "llama": 2,
-        "marian": 2,
-        "minicpm": 4,
-        "mistral": 2 if is_transformers_version(">=", "4.40.0") else 0,
-        "mixtral": 2 if is_transformers_version(">=", "4.40.0") else 0,
-        "mpt": 5,
-        "opt": 5,
-        "pegasus": 2,
-        "qwen": 2,
-        "phi": 2 if is_transformers_version(">=", "4.40.0") else 0,
-        "internlm2": 4,
-        "falcon": 2,
-        "falcon-40b": 2,
-        "persimmon": 2,
-        "biogpt": 5,
-        "aquila": 2,
-        "aquila2": 2,
-        "xverse": 2,
-        "internlm": 2,
-        "jais": 2,
-        "chatglm4": 6,
-        "decilm": 4,
-        "gemma": 1,
-        "olmo": 2,
-        "stablelm": 2,
-        "starcoder2": 2,
-        "dbrx": 2,
-        "cohere": 2,
-        "qwen2": 2,
-        "qwen2_moe": 4,
-        "arctic": 4,
-        "phi3": 2,
-        "gemma2": 4,
-        "exaone": 8,
-        "granite": 6,
-        "granite-moe": 6,
-        "glm": 28,
-        "mistral-nemo": 8,
-        "minicpm3": 6,
-        "phi3-moe": 2,
-        "deepseek": 2,
-        "opt_gptq": 12,
-        "mixtral_awq": 2,
-        "gemma3_text": 2,
-        "glm4": 2,
-        "qwen3": 2,
-        "qwen3_moe": 2,
-        "mamba": 0,
-        "falcon-mamba": 0,
-        "arcee": 2,
-        "ernie4_5": 2,
-    }
-
-    # TODO: remove gptq/awq from here
-    @parameterized.expand(SUPPORTED_ARCHITECTURES)
-    def test_compare_to_transformers(self, model_arch):
-        model_id = MODEL_NAMES[model_arch]
-
-        not_stateful = []
-        if is_openvino_version("<", "2024.0"):
-            not_stateful.append("mixtral")
-
-        if is_openvino_version("<", "2024.1"):
-            not_stateful.extend(["llama", "gemma", "gpt_bigcode"])
-
-        set_seed(SEED)
-
-        model_kwargs = {}
-        if model_arch in self.REMOTE_CODE_MODELS:
-            model_kwargs = {"trust_remote_code": True}
-
-        # starting from transformers 4.45.0 gemma2 uses eager attention by default, while ov - sdpa
-        if model_arch == "gemma2" and is_transformers_version(">=", "4.45.0"):
-            model_kwargs["attn_implementation"] = "sdpa"
-
-        ov_model = OVModelForCausalLM.from_pretrained(model_id, export=True, ov_config=F32_CONFIG, **model_kwargs)
-        self.assertIsInstance(ov_model.config, PretrainedConfig)
-        self.assertTrue(ov_model.use_cache)
-        tokenizer = AutoTokenizer.from_pretrained(model_id, trust_remote_code=model_arch in self.REMOTE_CODE_MODELS)
-        tokens = tokenizer("This is a sample output", return_tensors="pt")
-
-        ov_outputs = ov_model(**tokens)
-        self.assertTrue("logits" in ov_outputs)
-        self.assertIsInstance(ov_outputs.logits, torch.Tensor)
-        if model_arch in self.SUPPORTED_SSM_ARCHITECTURES:
-            from optimum.intel.openvino.modeling_decoder import OVMambaCache
-
-            self.assertTrue("cache_params" in ov_outputs)
-            self.assertIsInstance(ov_outputs.cache_params, OVMambaCache)
-            is_stateful = ov_model.config.model_type not in not_stateful
-            self.assertEqual(ov_model.stateful, is_stateful)
-            if is_stateful:
-                self.assertIsInstance(ov_outputs.cache_params.conv_states, list)
-                self.assertIsInstance(ov_outputs.cache_params.ssm_states, list)
-                self.assertTrue(
-                    len(ov_outputs.cache_params.conv_states) > 0 and len(ov_outputs.cache_params.ssm_states) > 0
-                )
-        else:
-            self.assertTrue("past_key_values" in ov_outputs)
-            self.assertIsInstance(ov_outputs.past_key_values, tuple)
-            is_stateful = ov_model.config.model_type not in not_stateful
-            self.assertEqual(ov_model.stateful, is_stateful)
-            if is_stateful:
-                self.assertTrue(len(ov_outputs.past_key_values) == 1 and len(ov_outputs.past_key_values[0]) == 0)
-
-        expected_num_sdpa = self.EXPECTED_NUM_SDPA.get(model_arch, 0)
-        num_sdpa = get_num_sdpa(ov_model.model)
-        self.assertEqual(
-            expected_num_sdpa,
-            num_sdpa,
-            f"Expected number of SDPA {expected_num_sdpa}, while model contains {num_sdpa}",
-        )
-
-        if "awq" in model_arch or "gptq" in model_arch:
-            model_kwargs["torch_dtype"] = torch.float32
-
-        set_seed(SEED)
-        with mock_torch_cuda_is_available("awq" in model_arch or "gptq" in model_arch):
-            transformers_model = AutoModelForCausalLM.from_pretrained(model_id, **model_kwargs)
-        if model_arch in ["qwen", "arctic", "chatglm4"]:
-            transformers_model.to(torch.float32)
-
-        with torch.no_grad():
-            with patch_awq_for_inference("awq" in model_arch):
-                transformers_outputs = transformers_model(**tokens)
-
-        # Compare tensor outputs
-        atol = 3e-3 if model_arch in ["minicpm", "qwen2-moe"] else 1e-4
-        # quantized models have different logits value range
-        if "awq" not in model_arch and "gptq" not in model_arch:
-            self.assertTrue(torch.allclose(ov_outputs.logits, transformers_outputs.logits, equal_nan=True, atol=atol))
-
-        # Qwen tokenizer does not support padding
-        if model_arch in ["qwen"]:
-            return
-
-        if model_arch not in ["chatglm", "chatglm4", "persimmon"]:
-            tokenizer.pad_token_id = tokenizer.eos_token_id
-
-        if model_arch == "persimmon":
-            tokenizer.pad_token_id = tokenizer.bos_token_id
-        # Compare batched generation
-        tokenizer.padding_side = "left"
-        tokens = tokenizer(["Today is a nice day and I am longer", "This is me"], return_tensors="pt", padding=True)
-        ov_model.generation_config.eos_token_id = None
-        transformers_model.generation_config.eos_token_id = None
-        ov_model.config.eos_token_id = None
-        transformers_model.config.eos_token_id = None
-        gen_config = GenerationConfig(
-            max_new_tokens=30,
-            min_new_tokens=30,
-            num_beams=1 if model_arch == "chatglm4" else 2,
-            do_sample=False,
-        )
-
-        ov_outputs = ov_model.generate(**tokens, generation_config=gen_config)
-
-        # TODO: add back once https://huggingface.co/katuni4ka/tiny-random-minicpm3/discussions/1 merged (for all models) as current mdoeling incompatible with transformers >= v4.49
-        if model_arch in {"deepseek"} and is_transformers_version(">=", "4.49"):
-            self.skipTest("Incompatible modeling code")
-
-        additional_inputs = {}
-        # gemma2 does not support dynamic cache, it is unfair to compare dynamic cache result vs hybrid cache,
-        # align cache representation in torch model
-        if model_arch in ["gemma2", "gemma3_text"]:
-            patch_update_causal_mask(transformers_model, "4.43.0")
-            transformers_model._supports_cache_class = True
-            transformers_model.generation_config.cache_implementation = None
-            from transformers.cache_utils import DynamicCache
-
-            additional_inputs = {"past_key_values": DynamicCache()}
-        with patch_awq_for_inference("awq" in model_arch):
-            transformers_outputs = transformers_model.generate(
-                **tokens, generation_config=gen_config, **additional_inputs
-            )
-
-        self.assertTrue(
-            torch.allclose(ov_outputs, transformers_outputs),
-            f"OV output {ov_outputs}\nTransformers output  {transformers_outputs}",
-        )
-
-        del transformers_model
-        del ov_model
-        gc.collect()
-
-    @parameterized.expand(SUPPORTED_ARCHITECTURES)
-    @pytest.mark.run_slow
-    @slow
-    def test_pipeline(self, model_arch):
-        set_seed(SEED)
-        model_kwargs = {}
-        model_id = MODEL_NAMES[model_arch]
-        if model_arch in self.REMOTE_CODE_MODELS:
-            model_kwargs = {
-                "config": AutoConfig.from_pretrained(model_id, trust_remote_code=True),
-                "trust_remote_code": True,
-            }
-        tokenizer = AutoTokenizer.from_pretrained(model_id, trust_remote_code=model_arch in self.REMOTE_CODE_MODELS)
-
-        if model_arch == "qwen":
-            tokenizer._convert_tokens_to_ids = lambda x: 0
-
-        additional_args = {}
-        if is_transformers_version(">=", "4.51"):
-            additional_args["use_model_defaults"] = False
-
-        set_seed(SEED)
-        model = OVModelForCausalLM.from_pretrained(model_id, use_cache=True, compile=False, **model_kwargs)
-        model.eval()
-        model.config.encoder_no_repeat_ngram_size = 0
-        model.to("cpu")
-        model.half()
-        model.compile()
-        pipe = pipeline("text-generation", model=model, tokenizer=tokenizer)
-        inputs = "My name is Arthur and I live in"
-        set_seed(SEED)
-        outputs = pipe(inputs, min_new_tokens=5, max_new_tokens=5, **additional_args, do_sample=False)
-        self.assertEqual(pipe.device, model.device)
-        self.assertTrue(all(inputs in item["generated_text"] for item in outputs))
-        ov_pipe = optimum_pipeline(
-            "text-generation",
-            model_id,
-            accelerator="openvino",
-            trust_remote_code=model_arch in self.REMOTE_CODE_MODELS,
-            tokenizer=tokenizer if model_arch == "qwen" else None,
-        )
-        set_seed(SEED)
-        ov_outputs = ov_pipe(inputs, min_new_tokens=5, max_new_tokens=5, **additional_args, do_sample=False)
-        self.assertEqual(outputs[-1]["generated_text"], ov_outputs[-1]["generated_text"])
-        del ov_pipe
-        del pipe
-        del model
-        gc.collect()
-
-    def test_model_and_decoder_same_device(self):
-        model_id = MODEL_NAMES["gpt2"]
-        model = OVModelForCausalLM.from_pretrained(model_id, export=True)
-        model.to("TEST")
-        self.assertEqual(model._device, "TEST")
-        # Verify that request is being reset
-        self.assertEqual(model.request, None)
-        del model
-        gc.collect()
-
-    def test_compare_with_and_without_past_key_values(self):
-        model_id = MODEL_NAMES["gpt2"]
-        tokenizer = AutoTokenizer.from_pretrained(model_id)
-        tokens = tokenizer("This is a sample input", return_tensors="pt")
-
-        model_with_pkv = OVModelForCausalLM.from_pretrained(model_id, export=True, use_cache=True, stateful=False)
-        outputs_model_with_pkv = model_with_pkv.generate(
-            **tokens, min_length=self.GENERATION_LENGTH, max_length=self.GENERATION_LENGTH, num_beams=1
-        )
-        del model_with_pkv
-
-        model_without_pkv = OVModelForCausalLM.from_pretrained(model_id, export=True, use_cache=False)
-        outputs_model_without_pkv = model_without_pkv.generate(
-            **tokens, min_length=self.GENERATION_LENGTH, max_length=self.GENERATION_LENGTH, num_beams=1
-        )
-        del model_without_pkv
-
-        self.assertTrue(torch.equal(outputs_model_with_pkv, outputs_model_without_pkv))
-        self.assertEqual(outputs_model_with_pkv.shape[1], self.GENERATION_LENGTH)
-        self.assertEqual(outputs_model_without_pkv.shape[1], self.GENERATION_LENGTH)
-
-        model_stateful = OVModelForCausalLM.from_pretrained(model_id, export=True, use_cache=True, stateful=True)
-        outputs_model_stateful = model_stateful.generate(
-            **tokens, min_length=self.GENERATION_LENGTH, max_length=self.GENERATION_LENGTH, num_beams=1
-        )
-        self.assertTrue(torch.equal(outputs_model_without_pkv, outputs_model_stateful))
-
-        logits = model_stateful(**tokens).logits
-        copy_logits = copy.deepcopy(logits)
-        tokens = tokenizer("Input sample", return_tensors="pt")
-        model_stateful(**tokens).logits
-        self.assertTrue(torch.equal(copy_logits, logits))
-        del model_stateful
-        gc.collect()
-
-    def test_print_model_properties(self):
-        # test setting OPENVINO_LOG_LEVEL to 3, which calls _print_compiled_model_properties
-        openvino_log_level = os.environ.get("OPENVINO_LOG_LEVEL", None)
-        os.environ["OPENVINO_LOG_LEVEL"] = "3"
-        model = OVModelForSequenceClassification.from_pretrained(MODEL_NAMES["bert"], export=True)
-        if openvino_log_level is not None:
-            os.environ["OPENVINO_LOG_LEVEL"] = openvino_log_level
-        # test calling function directly
-        _print_compiled_model_properties(model.request)
-
-    def test_auto_device_loading(self):
-        OV_MODEL_ID = "echarlaix/distilbert-base-uncased-finetuned-sst-2-english-openvino"
-        for device in ("AUTO", "AUTO:CPU"):
-            model = OVModelForSequenceClassification.from_pretrained(OV_MODEL_ID, device=device)
-            model.half()
-            self.assertEqual(model._device, device)
-            if device == "AUTO:CPU":
-                model = OVModelForSequenceClassification.from_pretrained(OV_MODEL_ID, device=device)
-                message = "Model should not be loaded from cache without explicitly setting CACHE_DIR"
-                self.assertFalse(model.request.get_property("LOADED_FROM_CACHE"), message)
-            del model
-            gc.collect()
-
-    def test_default_filling_attention_mask(self):
-        model_id = MODEL_NAMES["gpt2"]
-        model_with_cache = OVModelForCausalLM.from_pretrained(model_id, stateful=False, use_cache=True)
-        tokenizer = AutoTokenizer.from_pretrained(model_id)
-        tokenizer.pad_token = tokenizer.eos_token
-        texts = ["this is a simple input"]
-        tokens = tokenizer(texts, return_tensors="pt")
-        self.assertTrue("attention_mask" in model_with_cache.input_names)
-        outs = model_with_cache(**tokens)
-        attention_mask = tokens.pop("attention_mask")
-        outs_without_attn_mask = model_with_cache(**tokens)
-        self.assertTrue(torch.allclose(outs.logits, outs_without_attn_mask.logits))
-        input_ids = torch.argmax(outs.logits[:, -1:, :], dim=2)
-        past_key_values = outs.past_key_values
-        attention_mask = torch.ones((input_ids.shape[0], tokens.input_ids.shape[1] + 1), dtype=torch.long)
-        outs_step2 = model_with_cache(
-            input_ids=input_ids, attention_mask=attention_mask, past_key_values=past_key_values
-        )
-        outs_without_attn_mask_step2 = model_with_cache(input_ids=input_ids, past_key_values=past_key_values)
-        self.assertTrue(torch.allclose(outs_step2.logits, outs_without_attn_mask_step2.logits))
-        del model_with_cache
-        gc.collect()
-
-    def test_default_filling_attention_mask_and_position_ids(self):
-        model_id = MODEL_NAMES["llama"]
-        model_with_cache = OVModelForCausalLM.from_pretrained(model_id, stateful=False, use_cache=True)
-        tokenizer = AutoTokenizer.from_pretrained(model_id)
-        tokenizer.pad_token = tokenizer.eos_token
-        texts = ["this is a simple input"]
-        tokens = tokenizer(texts, return_tensors="pt")
-        self.assertTrue("position_ids" in model_with_cache.input_names)
-        outs = model_with_cache(**tokens)
-        attention_mask = tokens.pop("attention_mask")
-        outs_without_attn_mask = model_with_cache(**tokens)
-        self.assertTrue(torch.allclose(outs.logits, outs_without_attn_mask.logits))
-        input_ids = torch.argmax(outs.logits[:, -1:, :], dim=2)
-        past_key_values = outs.past_key_values
-        attention_mask = torch.ones((input_ids.shape[0], tokens.input_ids.shape[1] + 1), dtype=torch.long)
-        outs_step2 = model_with_cache(
-            input_ids=input_ids, attention_mask=attention_mask, past_key_values=past_key_values
-        )
-        outs_without_attn_mask_step2 = model_with_cache(input_ids=input_ids, past_key_values=past_key_values)
-        self.assertTrue(torch.allclose(outs_step2.logits, outs_without_attn_mask_step2.logits))
-        del model_with_cache
-        gc.collect()
-
-    @parameterized.expand(SUPPORTED_ARCHITECTURES)
-    @pytest.mark.run_slow
-    @slow
-    def test_beam_search(self, model_arch):
-        model_kwargs = {}
-        model_id = MODEL_NAMES[model_arch]
-        if model_arch in self.REMOTE_CODE_MODELS:
-            model_kwargs = {
-                "config": AutoConfig.from_pretrained(model_id, trust_remote_code=True),
-                "trust_remote_code": True,
-            }
-
-        # starting from transformers 4.45.0 gemma2 uses eager attention by default, while ov - sdpa
-        if model_arch == "gemma2" and is_transformers_version(">=", "4.45.0"):
-            model_kwargs["attn_implementation"] = "sdpa"
-
-        # Qwen tokenizer does not support padding, chatglm, glm4 testing models produce nan that incompatible with beam search
-        if model_arch in ["qwen", "chatglm", "chatglm4"]:
-            return
-
-        # TODO: add back once https://huggingface.co/katuni4ka/tiny-random-minicpm3/discussions/1 merged (for all models) as current mdoeling incompatible with transformers >= v4.49
-        if model_arch in {"deepseek"} and is_transformers_version(">=", "4.49"):
-            self.skipTest("Incompatible modeling code")
-
-        tokenizer = AutoTokenizer.from_pretrained(model_id, trust_remote_code=model_arch in self.REMOTE_CODE_MODELS)
-        if model_arch == "persimmon":
-            tokenizer.pad_token_id = tokenizer.bos_token_id
-            tokenizer.eos_token_id = tokenizer.bos_token_id
-
-        beam_search_gen_config = GenerationConfig(
-            max_new_tokens=10,
-            min_new_tokens=10,
-            num_beams=4,
-            do_sample=False,
-            eos_token_id=None,
-        )
-
-        beam_sample_gen_config = GenerationConfig(
-            max_new_tokens=10,
-            min_new_tokens=10,
-            num_beams=4,
-            do_sample=True,
-            eos_token_id=None,
-        )
-
-        if model_arch in ["minicpm", "internlm2"]:
-            beam_sample_gen_config.top_k = 1
-
-        group_beam_search_gen_config = GenerationConfig(
-            max_new_tokens=10,
-            min_new_tokens=10,
-            num_beams=4,
-            do_sample=False,
-            eos_token_id=None,
-            num_beam_groups=2,
-            diversity_penalty=0.0000001,
-        )
-        force_word = "cat"
-        force_words_ids = [tokenizer([force_word], add_special_tokens=False).input_ids]
-        constrained_beam_search_gen_config = GenerationConfig(
-            max_new_tokens=10,
-            min_new_tokens=10,
-            num_beams=4,
-            do_sample=False,
-            eos_token_id=None,
-            force_words_ids=force_words_ids,
-        )
-
-        gen_configs = [
-            beam_search_gen_config,
-            beam_sample_gen_config,
-            group_beam_search_gen_config,
-            constrained_beam_search_gen_config,
-        ]
-        set_seed(SEED)
-        ov_model_stateful = OVModelForCausalLM.from_pretrained(
-            model_id, export=True, use_cache=True, stateful=True, **model_kwargs
-        )
-        set_seed(SEED)
-        ov_model_stateless = OVModelForCausalLM.from_pretrained(
-            model_id, export=True, use_cache=True, stateful=False, **model_kwargs
-        )
-        if "awq" in model_arch or "gptq" in model_arch:
-            # infer in FP32
-            model_kwargs["torch_dtype"] = torch.float32
-
-        set_seed(SEED)
-        with mock_torch_cuda_is_available("awq" in model_arch or "gptq" in model_arch):
-            transformers_model = AutoModelForCausalLM.from_pretrained(model_id, **model_kwargs)
-        if model_arch == "arctic":
-            transformers_model.to(torch.float32)
-        additional_inputs = {}
-        # gemma2 does not support dynamic cache, it is unfair to compare dynamic cache result vs hybrid cache, align cache representation in torch model
-        if model_arch in ["gemma2", "gemma3_text"]:
-            patch_update_causal_mask(transformers_model, "4.43.0")
-            transformers_model._supports_cache_class = True
-            transformers_model.generation_config.cache_implementation = None
-        tokenizer.pad_token_id = tokenizer.eos_token_id
-        tokenization_args = {}
-        if is_transformers_version(">=", "4.45") and model_arch == "gpt_neo":
-            tokenization_args["padding_side"] = "left"
-        tokens = tokenizer(
-            ["Today is a nice day and I am longer", "This is me"],
-            return_tensors="pt",
-            padding=True,
-            **tokenization_args,
-        )
-        ov_model_stateful.generation_config.eos_token_id = None
-        ov_model_stateful.generation_config.forced_eos_token_id = None
-        ov_model_stateful.generation_config.encoder_no_repeat_ngram_size = None
-        ov_model_stateful.generation_config.do_sample = False
-        ov_model_stateless.generation_config.eos_token_id = None
-        ov_model_stateless.generation_config.forced_eos_token_id = None
-        ov_model_stateless.generation_config.encoder_no_repeat_ngram_size = None
-        ov_model_stateless.generation_config.do_sample = False
-        transformers_model.generation_config.eos_token_id = None
-        transformers_model.generation_config.forced_eos_token_id = None
-        transformers_model.generation_config.encoder_no_repeat_ngram_size = None
-        transformers_model.generation_config.do_sample = False
-        ov_model_stateful.config.eos_token_id = None
-        ov_model_stateless.config.eos_token_id = None
-        transformers_model.config.eos_token_id = None
-
-        if is_transformers_version(">=", "4.51"):
-            additional_inputs["use_model_defaults"] = False
-
-        for gen_config in gen_configs:
-            if gen_config.do_sample and model_arch in ["baichuan2-13b", "olmo"]:
-                continue
-            if gen_config.num_beams > 1 and is_transformers_version(">=", "4.51.0") and model_arch in ["mixtral_awq"]:
-                continue
-            set_seed(SEED)
-
-            if model_arch in ["gemma2", "gemma3_text"]:
-                from transformers.cache_utils import DynamicCache
-
-                additional_inputs["past_key_values"] = DynamicCache()
-            with patch_awq_for_inference("awq" in model_arch):
-                transformers_outputs = transformers_model.generate(
-                    **tokens, generation_config=gen_config, **additional_inputs
-                )
-            set_seed(SEED)
-            ov_stateful_outputs = ov_model_stateful.generate(**tokens, generation_config=gen_config)
-            self.assertTrue(
-                torch.equal(ov_stateful_outputs, transformers_outputs),
-                f"generation config : {gen_config}, transformers output {transformers_outputs}, ov_model_stateful output {ov_stateful_outputs}",
-            )
-
-            set_seed(SEED)
-            ov_stateless_outputs = ov_model_stateless.generate(**tokens, generation_config=gen_config)
-            self.assertTrue(
-                torch.equal(ov_stateless_outputs, transformers_outputs),
-                f"generation config : {gen_config}, transformers output {transformers_outputs}, ov_model_stateless output {ov_stateless_outputs}",
-            )
-
-    def test_load_with_different_dtype(self):
-        set_seed(SEED)
-        model_id = MODEL_NAMES["llama"]
-        pt_model = AutoModelForCausalLM.from_pretrained(
-            model_id,
-        )
-        tokenizer = AutoTokenizer.from_pretrained(model_id)
-
-        texts = ["this is a simple input"]
-        test_input = tokenizer(texts, return_tensors="pt")
-
-        ref_logits = pt_model(**test_input).logits
-        torch_dtypes = [None, "auto", "float32", torch.float16]
-        if is_openvino_version(">", "2024.2.0"):
-            torch_dtypes.append("bfloat16")
-
-        for dtype in torch_dtypes:
-            ov_model = OVModelForCausalLM.from_pretrained(model_id=model_id, export=True, torch_dtype=dtype)
-            ov_logits = ov_model(**test_input).logits
-            self.assertTrue(
-                torch.allclose(torch.Tensor(ov_logits), ref_logits, atol=5e-3),
-                f"values are not close for {dtype if dtype is not None else 'None'}, max diff = {torch.abs(ov_logits - ref_logits).max()}",
-            )
-
-
 class OVModelForMaskedLMIntegrationTest(unittest.TestCase):
     SUPPORTED_ARCHITECTURES = (
         "albert",
@@ -1931,184 +1251,6 @@ class OVModelForImageClassificationIntegrationTest(unittest.TestCase):
             ov_model.save_pretrained(model_save_path)
             model = OVModelForImageClassification.from_pretrained(model_save_path)
             model(pixel_values=torch.zeros((5, 3, model.config.image_size, model.config.image_size)))
-        gc.collect()
-
-
-class OVModelForSeq2SeqLMIntegrationTest(unittest.TestCase):
-    SUPPORTED_ARCHITECTURES = (
-        "bart",
-        # "bigbird_pegasus",
-        "blenderbot",
-        "blenderbot-small",
-        # "longt5",
-        "m2m_100",
-        "marian",
-        "mbart",
-        "mt5",
-        "pegasus",
-        "t5",
-    )
-
-    GENERATION_LENGTH = 100
-    SPEEDUP_CACHE = 1.1
-
-    SUPPORT_STATEFUL = ("t5", "mt5")
-    if is_transformers_version(">=", "4.52.0"):
-        SUPPORT_STATEFUL += ("bart", "blenderbot", "blenderbot-small", "m2m_100", "marian", "mbart")
-    if is_transformers_version(">=", "4.53.0"):
-        SUPPORT_STATEFUL += ("pegasus",)
-
-    @parameterized.expand(SUPPORTED_ARCHITECTURES)
-    def test_compare_to_transformers(self, model_arch):
-        model_id = MODEL_NAMES[model_arch]
-        set_seed(SEED)
-        ov_model = OVModelForSeq2SeqLM.from_pretrained(model_id, export=True, ov_config=F32_CONFIG)
-        ov_stateless_model = OVModelForSeq2SeqLM.from_pretrained(
-            model_id, export=True, use_cache=False, stateful=False, ov_config=F32_CONFIG
-        )
-        expected_stateful = is_transformers_version(">", "4.43") and model_arch in self.SUPPORT_STATEFUL
-        self.assertEqual(ov_model.decoder.stateful, expected_stateful)
-        self.assertEqual(model_has_state(ov_model.decoder.model), expected_stateful)
-        check_with_past_available = self.assertIsNone if expected_stateful else self.assertIsNotNone
-        check_with_past_available(ov_model.decoder_with_past)
-        self.assertIsInstance(ov_model.encoder, OVEncoder)
-        self.assertIsInstance(ov_model.decoder, OVDecoder)
-        if not ov_model.decoder.stateful:
-            self.assertIsInstance(ov_model.decoder_with_past, OVDecoder)
-            self.assertIsInstance(ov_model.config, PretrainedConfig)
-
-        transformers_model = AutoModelForSeq2SeqLM.from_pretrained(model_id)
-        tokenizer = AutoTokenizer.from_pretrained(model_id)
-        tokens = tokenizer("This is a sample input", return_tensors="pt")
-        decoder_start_token_id = transformers_model.config.decoder_start_token_id if model_arch != "mbart" else 2
-        decoder_inputs = {"decoder_input_ids": torch.ones((1, 1), dtype=torch.long) * decoder_start_token_id}
-        ov_outputs = ov_model(**tokens, **decoder_inputs)
-        ov_stateless_outputs = ov_stateless_model(**tokens, **decoder_inputs)
-
-        self.assertTrue("logits" in ov_outputs)
-        self.assertIsInstance(ov_outputs.logits, torch.Tensor)
-
-        with torch.no_grad():
-            transformers_outputs = transformers_model(**tokens, **decoder_inputs)
-        # Compare tensor outputs
-        self.assertTrue(torch.allclose(ov_outputs.logits, transformers_outputs.logits, atol=5e-3))
-        self.assertTrue(torch.allclose(ov_stateless_outputs.logits, transformers_outputs.logits, atol=5e-3))
-        gen_config = GenerationConfig(
-            max_new_tokens=10,
-            min_new_tokens=10,
-            num_beams=2,
-            do_sample=False,
-            eos_token_id=None,
-        )
-
-        set_seed(SEED)
-        generated_tokens = transformers_model.generate(**tokens, generation_config=gen_config)
-        set_seed(SEED)
-        ov_generated_tokens = ov_model.generate(**tokens, generation_config=gen_config)
-        set_seed(SEED)
-        ov_stateless_generated_tokens = ov_stateless_model.generate(**tokens, generation_config=gen_config)
-
-        self.assertTrue(torch.equal(generated_tokens, ov_generated_tokens))
-        self.assertTrue(torch.equal(generated_tokens, ov_stateless_generated_tokens))
-
-        del transformers_model
-        del ov_model
-
-        gc.collect()
-
-    @parameterized.expand(SUPPORTED_ARCHITECTURES)
-    @pytest.mark.run_slow
-    @slow
-    def test_pipeline(self, model_arch):
-        set_seed(SEED)
-        model_id = MODEL_NAMES[model_arch]
-        tokenizer = AutoTokenizer.from_pretrained(model_id)
-        inputs = "This is a test"
-        model = OVModelForSeq2SeqLM.from_pretrained(model_id, compile=False)
-        model.eval()
-        model.half()
-        model.to("cpu")
-        model.compile()
-
-        # Summarization
-        pipe = pipeline("summarization", model=model, tokenizer=tokenizer)
-        outputs = pipe(inputs)
-        self.assertEqual(pipe.device, model.device)
-        self.assertIsInstance(outputs[0]["summary_text"], str)
-
-        # Translation
-        pipe = pipeline("translation_en_to_fr", model=model, tokenizer=tokenizer)
-        outputs = pipe(inputs)
-        self.assertEqual(pipe.device, model.device)
-        self.assertIsInstance(outputs[0]["translation_text"], str)
-
-        # Text2Text generation
-        pipe = pipeline("text2text-generation", model=model, tokenizer=tokenizer)
-        outputs = pipe(inputs, max_new_tokens=20)
-        self.assertEqual(pipe.device, model.device)
-        self.assertIsInstance(outputs[0]["generated_text"], str)
-
-        ov_pipe = optimum_pipeline("text2text-generation", model_id, accelerator="openvino")
-        ov_outputs = ov_pipe(inputs, max_new_tokens=20)
-        self.assertEqual(outputs[-1]["generated_text"], ov_outputs[-1]["generated_text"])
-        del ov_pipe
-        del pipe
-        del model
-        gc.collect()
-
-    @parameterized.expand(SUPPORTED_ARCHITECTURES)
-    @pytest.mark.run_slow
-    @slow
-    def test_generate_utils(self, model_arch):
-        model_id = MODEL_NAMES[model_arch]
-        model = OVModelForSeq2SeqLM.from_pretrained(model_id, export=True)
-        tokenizer = AutoTokenizer.from_pretrained(model_id)
-        text = "This is a sample input"
-        tokens = tokenizer(text, return_tensors="pt")
-
-        # General case
-        outputs = model.generate(**tokens)
-        outputs = tokenizer.batch_decode(outputs, skip_special_tokens=True)
-        self.assertIsInstance(outputs[0], str)
-
-        # With input ids
-        outputs = model.generate(input_ids=tokens["input_ids"])
-        outputs = tokenizer.batch_decode(outputs, skip_special_tokens=True)
-        self.assertIsInstance(outputs[0], str)
-        del model
-
-        gc.collect()
-
-    def test_compare_with_and_without_past_key_values(self):
-        model_id = MODEL_NAMES["bart"]
-        tokenizer = AutoTokenizer.from_pretrained(model_id)
-        text = "This is a sample input"
-        tokens = tokenizer(text, return_tensors="pt")
-
-        model_with_pkv = OVModelForSeq2SeqLM.from_pretrained(model_id, export=True, use_cache=True)
-        _ = model_with_pkv.generate(**tokens)  # warmup
-        with Timer() as with_pkv_timer:
-            outputs_model_with_pkv = model_with_pkv.generate(
-                **tokens, min_length=self.GENERATION_LENGTH, max_length=self.GENERATION_LENGTH, num_beams=1
-            )
-
-        model_without_pkv = OVModelForSeq2SeqLM.from_pretrained(model_id, export=True, use_cache=False)
-        _ = model_without_pkv.generate(**tokens)  # warmup
-        with Timer() as without_pkv_timer:
-            outputs_model_without_pkv = model_without_pkv.generate(
-                **tokens, min_length=self.GENERATION_LENGTH, max_length=self.GENERATION_LENGTH, num_beams=1
-            )
-
-        self.assertTrue(torch.equal(outputs_model_with_pkv, outputs_model_without_pkv))
-        self.assertEqual(outputs_model_with_pkv.shape[1], self.GENERATION_LENGTH)
-        self.assertEqual(outputs_model_without_pkv.shape[1], self.GENERATION_LENGTH)
-        self.assertTrue(
-            without_pkv_timer.elapsed / with_pkv_timer.elapsed > self.SPEEDUP_CACHE,
-            f"With pkv latency: {with_pkv_timer.elapsed:.3f} ms, without pkv latency: {without_pkv_timer.elapsed:.3f} ms,"
-            f" speedup: {without_pkv_timer.elapsed / with_pkv_timer.elapsed:.3f}",
-        )
-        del model_with_pkv
-        del model_without_pkv
         gc.collect()
 
 
@@ -2843,221 +1985,6 @@ class OVModelForVisualCausalLMIntegrationTest(unittest.TestCase):
                 save_dir, compile=False, trust_remote_code=model_arch in self.REMOTE_CODE_MODELS
             )
             self.assertIsInstance(ov_restored_model, type(ov_model))
-
-
-class OVModelForSpeechSeq2SeqIntegrationTest(unittest.TestCase):
-    SUPPORTED_ARCHITECTURES = ("whisper",)
-
-    def _generate_random_audio_data(self):
-        np.random.seed(10)
-        t = np.linspace(0, 5.0, int(5.0 * 22050), endpoint=False)
-        # generate pure sine wave at 220 Hz
-        audio_data = 0.5 * np.sin(2 * np.pi * 220 * t)
-        return audio_data
-
-    @parameterized.expand(SUPPORTED_ARCHITECTURES)
-    def test_compare_to_transformers(self, model_arch):
-        set_seed(SEED)
-        model_id = MODEL_NAMES[model_arch]
-        transformers_model = AutoModelForSpeechSeq2Seq.from_pretrained(model_id)
-        ov_model = OVModelForSpeechSeq2Seq.from_pretrained(model_id, export=True, ov_config=F32_CONFIG)
-        ov_model_stateless = OVModelForSpeechSeq2Seq.from_pretrained(
-            model_id, export=True, ov_config=F32_CONFIG, stateful=False
-        )
-        self.assertIsInstance(ov_model.config, PretrainedConfig)
-        # whisper cache class support implemented in 4.43
-        expected_stateful = is_transformers_version(">", "4.43")
-        self.assertEqual(ov_model.decoder.stateful, expected_stateful)
-        self.assertEqual(model_has_state(ov_model.decoder.model), expected_stateful)
-        check_with_past_available = self.assertIsNone if expected_stateful else self.assertIsNotNone
-        check_with_past_available(ov_model.decoder_with_past)
-
-        processor = get_preprocessor(model_id)
-        data = self._generate_random_audio_data()
-        pt_features = processor.feature_extractor(data, return_tensors="pt")
-        decoder_start_token_id = transformers_model.config.decoder_start_token_id
-        decoder_inputs = {"decoder_input_ids": torch.ones((1, 1), dtype=torch.long) * decoder_start_token_id}
-
-        with torch.no_grad():
-            transformers_outputs = transformers_model(**pt_features, **decoder_inputs)
-
-        for input_type in ["pt", "np"]:
-            features = processor.feature_extractor(data, return_tensors=input_type)
-
-            if input_type == "np":
-                decoder_inputs = {"decoder_input_ids": np.ones((1, 1), dtype=np.int64) * decoder_start_token_id}
-
-            ov_outputs = ov_model(**features, **decoder_inputs)
-            ov_stateless_outputs = ov_model_stateless(**features, **decoder_inputs)
-            self.assertIn("logits", ov_outputs)
-            # Compare tensor outputs
-            self.assertTrue(torch.allclose(torch.Tensor(ov_outputs.logits), transformers_outputs.logits, atol=1e-3))
-            self.assertTrue(
-                torch.allclose(torch.Tensor(ov_stateless_outputs.logits), transformers_outputs.logits, atol=1e-3)
-            )
-
-        generate_kwrgs = {}
-        if is_transformers_version(">=", "4.50"):
-            generate_kwrgs = {"use_model_defaults": False}
-
-        gen_config = GenerationConfig(
-            max_new_tokens=10,
-            min_new_tokens=10,
-            num_beams=2,
-            do_sample=False,
-            eos_token_id=None,
-        )
-
-        set_seed(SEED)
-        generated_tokens = transformers_model.generate(**pt_features, generation_config=gen_config, **generate_kwrgs)
-        set_seed(SEED)
-        ov_generated_tokens = ov_model.generate(**pt_features, generation_config=gen_config, **generate_kwrgs)
-        set_seed(SEED)
-        ov_stateless_generated_tokens = ov_model_stateless.generate(
-            **pt_features, generation_config=gen_config, **generate_kwrgs
-        )
-
-        self.assertTrue(torch.equal(generated_tokens, ov_generated_tokens))
-        self.assertTrue(torch.equal(generated_tokens, ov_stateless_generated_tokens))
-
-        del transformers_model
-        del ov_model
-        gc.collect()
-
-    @parameterized.expand(SUPPORTED_ARCHITECTURES)
-    @pytest.mark.run_slow
-    @slow
-    def test_pipeline(self, model_arch):
-        set_seed(SEED)
-        model_id = MODEL_NAMES[model_arch]
-        model = OVModelForSpeechSeq2Seq.from_pretrained(model_id)
-        processor = get_preprocessor(model_id)
-        pipe = pipeline(
-            "automatic-speech-recognition",
-            model=model,
-            tokenizer=processor.tokenizer,
-            feature_extractor=processor.feature_extractor,
-        )
-        inputs = self._generate_random_audio_data()
-        outputs = pipe(inputs)
-        self.assertIsInstance(outputs["text"], str)
-
-        ov_pipe = optimum_pipeline("automatic-speech-recognition", model_id, accelerator="openvino")
-        ov_outputs = ov_pipe(inputs)
-        self.assertEqual(outputs["text"], ov_outputs["text"])
-        del ov_pipe
-        del pipe
-        del model
-        gc.collect()
-
-
-class OVModelForVision2SeqIntegrationTest(unittest.TestCase):
-    SUPPORTED_ARCHITECTURES = ["vision-encoder-decoder", "trocr", "donut"]
-
-    GENERATION_LENGTH = 100
-    SPEEDUP_CACHE = 1.1
-
-    def _get_sample_image(self):
-        url = TEST_IMAGE_URL
-        image = Image.open(requests.get(url, stream=True).raw)
-        return image
-
-    def _get_preprocessors(self, model_id):
-        image_processor = AutoImageProcessor.from_pretrained(model_id)
-        tokenizer = AutoTokenizer.from_pretrained(model_id)
-
-        return image_processor, tokenizer
-
-    def test_load_vanilla_transformers_which_is_not_supported(self):
-        with self.assertRaises(Exception) as context:
-            _ = OVModelForVision2Seq.from_pretrained(MODEL_NAMES["bert"], export=True)
-
-        self.assertIn("only supports the tasks", str(context.exception))
-
-    @parameterized.expand(SUPPORTED_ARCHITECTURES)
-    @pytest.mark.run_slow
-    @slow
-    def test_generate_utils(self, model_arch: str):
-        model_id = MODEL_NAMES[model_arch]
-        model = OVModelForVision2Seq.from_pretrained(model_id, export=True)
-        feature_extractor, tokenizer = self._get_preprocessors(model_id)
-
-        data = self._get_sample_image()
-        features = feature_extractor(data, return_tensors="pt")
-
-        outputs = model.generate(inputs=features["pixel_values"])
-        res = tokenizer.batch_decode(outputs, skip_special_tokens=True)
-        self.assertIsInstance(res[0], str)
-
-        gc.collect()
-
-    @parameterized.expand(SUPPORTED_ARCHITECTURES)
-    def test_compare_to_transformers(self, model_arch: str):
-        model_id = MODEL_NAMES[model_arch]
-        ov_model = OVModelForVision2Seq.from_pretrained(model_id, export=True)
-
-        self.assertIsInstance(ov_model.encoder, OVEncoder)
-
-        self.assertIsInstance(ov_model.decoder, OVDecoder)
-        self.assertIsInstance(ov_model.decoder_with_past, OVDecoder)
-
-        self.assertIsInstance(ov_model.config, PretrainedConfig)
-
-        set_seed(SEED)
-        transformers_model = AutoModelForVision2Seq.from_pretrained(model_id)
-        feature_extractor, tokenizer = self._get_preprocessors(model_id)
-
-        data = self._get_sample_image()
-
-        start_token = "<s>"
-        decoder_start_token_id = tokenizer.encode(start_token)[0]
-
-        extra_inputs = [{}, {}]
-
-        for extra_inps in extra_inputs:
-            features = feature_extractor(data, return_tensors="pt")
-            decoder_inputs = {"decoder_input_ids": torch.ones((1, 1), dtype=torch.long) * decoder_start_token_id}
-
-            with torch.no_grad():
-                transformers_outputs = transformers_model(**features, **decoder_inputs, **extra_inps, use_cache=True)
-            input_type = "pt"
-            features = feature_extractor(data, return_tensors=input_type)
-            ov_outputs = ov_model(**features, **decoder_inputs, **extra_inps)
-
-            self.assertTrue("logits" in ov_outputs)
-
-            # Compare tensor outputs
-            self.assertTrue(torch.allclose(torch.Tensor(ov_outputs.logits), transformers_outputs.logits, atol=1e-3))
-
-        gc.collect()
-
-    @parameterized.expand(SUPPORTED_ARCHITECTURES)
-    @pytest.mark.run_slow
-    @slow
-    def test_pipeline(self, model_arch: str):
-        set_seed(SEED)
-        model_id = MODEL_NAMES[model_arch]
-        ov_model = OVModelForVision2Seq.from_pretrained(model_id, compile=False)
-        feature_extractor, tokenizer = self._get_preprocessors(model_id)
-        ov_model.reshape(1, -1)
-        ov_model.compile()
-
-        # Image caption generation
-        pipe = pipeline(
-            "image-to-text",
-            model=ov_model,
-            tokenizer=tokenizer,
-            feature_extractor=feature_extractor,
-        )
-        inputs = self._get_sample_image()
-        outputs = pipe(inputs, max_new_tokens=3)
-        self.assertEqual(pipe.device, ov_model.device)
-        self.assertIsInstance(outputs[0]["generated_text"], str)
-        ov_pipe = optimum_pipeline("image-to-text", model_id, accelerator="openvino")
-        ov_outputs = ov_pipe(inputs, max_new_tokens=3)
-        self.assertEqual(outputs[-1]["generated_text"], ov_outputs[-1]["generated_text"])
-
-        gc.collect()
 
 
 class OVModelForCustomTasksIntegrationTest(unittest.TestCase):
